backend/requirements.txt
Django==5.0
djangorestframework
django-cors-headers
psycopg2-binary==2.9.9
requests
# Add other dependencies as needed
